짤게시판 리팩토링 - 1st

+++ src/main/java/com/example/backend/image/ImageController.java
@@ -20,11 +20,14 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/boards/image")
@@ -38,30 +41,35 @@ public class ImageController {
     private final AwsS3Service awsS3Service;
 
     @PostMapping(value = "",consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ResponseEntity<String> createImagePost(@RequestPart ImageBoardDto imageBoardDto, @RequestPart List<MultipartFile> files){
-        Long loginId = 1L;
-        User user = userService.findUserById(loginId);
-        System.out.println(imageBoardDto);
+    public ResponseEntity<String> createImagePost(@RequestPart ImageBoardDto imageBoardDto, @RequestPart MultipartFile imageFile){
 
-        String url = awsS3Service.uploadFile(files.get(0));
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = ((UserDetails) principal).getUsername();
+        User user = userService.findUserById(Long.parseLong(username));
+
+        String url = awsS3Service.uploadFile(imageFile);
         List<String> hashtags = imageBoardDto.getHashtags();
         ImageBoard imageBoard = ImageBoard.builder()
                 .imageBoardTitle(imageBoardDto.getImageBoardTitle()).imageUrl(url).createdTime(LocalDateTime.now()).user(user)
                 .build();
-        imageBoardService.savePost(imageBoard,files);
+        imageBoardService.savePost(imageBoard);
         imageBoardService.saveImageBoardHashtag(hashtags, imageBoard);
         return new ResponseEntity<>("짤 게시판 등록 성공", HttpStatus.CREATED);
     }
 
     @GetMapping("/{imageBoardId}")
     public ResponseEntity<DetailImageBoardResponse> getDetailImageBoard(@PathVariable Long imageBoardId){
-        Long loginId = 1L;
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = ((UserDetails) principal).getUsername();
+        User user = userService.findUserById(Long.parseLong(username));
+
         ImageBoard targetImageBoard = imageBoardService.findImageBoardEntity(imageBoardId);
         if(targetImageBoard == null){
             return new ResponseEntity<>(null,HttpStatus.NOT_FOUND);
         }
         imageBoardService.updateImageBoardViewCount(targetImageBoard);
-        DetailImageBoardResponse imageBoardResponse = imageBoardService.getDetailImageResponse(targetImageBoard, loginId);
+        DetailImageBoardResponse imageBoardResponse = imageBoardService.getDetailImageResponse(targetImageBoard, user.getUserId());
         return new ResponseEntity<>(imageBoardResponse, HttpStatus.OK);
     }
 
@@ -78,9 +86,12 @@ public class ImageController {
 
     @PostMapping("/{imageBoardId}/likes")
     public ResponseEntity<Integer> setLikeOnImageBoard(@PathVariable Long imageBoardId){
-        Long loginId = 1L;
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = ((UserDetails) principal).getUsername();
+        User user = userService.findUserById(Long.parseLong(username));
+
         ImageBoard imageBoard = imageBoardService.findImageBoardEntity(imageBoardId);
-        User user = userService.findUserById(loginId);
         Like like = likeService.findExistingLike(imageBoard,user);
         Integer totalLikeCount;
         if(like==null){
@@ -99,9 +110,12 @@ public class ImageController {
 
     @PostMapping("/{imageBoardId}/reports")
     public ResponseEntity<String> reportImageBoard(@PathVariable Long imageBoardId){
-        Long loginId = 5L;
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = ((UserDetails) principal).getUsername();
+        User user = userService.findUserById(Long.parseLong(username));
+
         ImageBoard imageBoard = imageBoardService.findImageBoardEntity(imageBoardId);
-        User user = userService.findUserById(loginId);
         if(reportService.checkReportExistence(user,imageBoard)){
             return new ResponseEntity<>("해당 사용자가 이미 신고한 짤",HttpStatus.OK);
         }
@@ -116,8 +130,17 @@ public class ImageController {
 
     @DeleteMapping("/{imageBoardId}")
     public ResponseEntity<String> deleteImageBoard(@PathVariable Long imageBoardId){
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = ((UserDetails) principal).getUsername();
+        User user = userService.findUserById(Long.parseLong(username));
+
         ImageBoard imageBoard = imageBoardService.findImageBoardEntity(imageBoardId);
-        imageBoardService.deleteImageBoard(imageBoard);
-        return new ResponseEntity<>("짤 게시글 삭제 성공",HttpStatus.NO_CONTENT);
+
+        if(Objects.equals(imageBoard.getUser().getUserId(), user.getUserId())){
+            imageBoardService.deleteImageBoard(imageBoard);
+            return new ResponseEntity<>("짤 게시글 삭제 성공",HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>("작성자만 삭제 가능", HttpStatus.BAD_REQUEST);
     }
 }
